refactor Action.kt into coroutines package
