+++ src/main/kotlin/ca/warp7/frc/coroutines/Action.kt
@@ -1,14 +1,11 @@
-package ca.warp7.frc
+package ca.warp7.frc.coroutines
 
-import ca.warp7.frc.CycleState.*
+import ca.warp7.frc.coroutines.CycleState.*
 //import edu.wpi.first.wpilibj.Notifier
 //import edu.wpi.first.wpilibj.RobotBase
 //import edu.wpi.first.wpilibj.Timer
 import java.lang.StringBuilder
 import kotlin.coroutines.*
-
-@Experimental
-annotation class ExperimentalActionDSL
 
 /**
  * An [Action] defines any self contained action that can be executed by the robot.
@@ -239,175 +236,3 @@ open class Action {
     }
 }
 
-@ExperimentalActionDSL
-class SynchronizedControl {
-    private var currentAction = Action()
-
-    fun setAction(action: Action) {
-        synchronized(this) {
-            currentAction = action
-            currentAction.advanceState()
-        }
-    }
-
-    fun interrupt() {
-        synchronized(this) {
-            currentAction.stop()
-        }
-    }
-
-    fun updateActions() {
-        synchronized(this) {
-            currentAction.advanceState()
-        }
-    }
-}
-
-private enum class CycleState {
-    FirstCycle, Periodic, Done, Idle
-}
-
-private enum class CoroutineState {
-    Ready, Done
-}
-
-/**
- * [ActionDSL] marks a method to be part of the Action library
- */
-
-@DslMarker
-@ExperimentalActionDSL
-annotation class ActionDSL
-
-@ExperimentalActionDSL
-@ActionDSL
-interface ActionCoroutine {
-
-    @ActionDSL
-    suspend operator fun <T : Action> T.unaryPlus()
-
-    @ActionDSL
-    suspend fun cancel()
-
-    @ActionDSL
-    suspend fun delay(seconds: Number)
-
-    @ActionDSL
-    suspend fun lock()
-
-    @ActionDSL
-    suspend fun free()
-
-    @ActionDSL
-    suspend fun skip()
-
-    @ActionDSL
-    suspend fun parallel(block: suspend ActionCoroutine.() -> Unit)
-}
-
-@ExperimentalActionDSL
-private class CoroutineWithContinuation(
-        val handle: Int,
-        val action: Action,
-        val debug: Boolean = false
-) : ActionCoroutine, Continuation<Unit> {
-
-    var nextStep: Continuation<Unit>? = null
-
-    var locked = false
-
-    var state = CoroutineState.Ready
-
-    val subActions: MutableList<Action> = mutableListOf()
-
-    override fun resumeWith(result: Result<Unit>) {
-        result.getOrThrow()
-    }
-
-    override val context: CoroutineContext
-        get() = EmptyCoroutineContext
-
-    override fun toString(): String {
-        return ""
-    }
-
-    fun advanceStateIsDone(): Boolean {
-        if (state == CoroutineState.Ready) {
-            val next = nextStep
-            if (next == null) {
-                state = CoroutineState.Done
-                return true
-            } else {
-                subActions.forEach {
-                    it.advanceState()
-                }
-                nextStep = null // Keep track of state using null
-                next.resume(Unit)
-            }
-        }
-        return false
-    }
-
-    override suspend fun <T : Action> T.unaryPlus() {
-        if (locked) {
-            subActions.add(this)
-        } else {
-            subActions.add(this)
-        }
-    }
-
-    override suspend fun cancel() {
-        for (subAction in subActions) {
-            subAction.interrupt()
-        }
-    }
-
-    override suspend fun delay(seconds: Number) {
-        TODO()
-    }
-
-    override suspend fun lock() {
-        locked = true
-    }
-
-    override suspend fun free() {
-        locked = false
-    }
-
-    override suspend fun skip() {
-        suspendCoroutine<Unit> { continuation ->
-            nextStep = continuation
-        }
-    }
-
-    override suspend fun parallel(block: suspend ActionCoroutine.() -> Unit) {
-        action.runRoutine(debug, block)
-    }
-}
-
-@ExperimentalActionDSL
-@ActionDSL
-fun routineOf(debug: Boolean = false, block: suspend ActionCoroutine.() -> Unit) = Routine(debug, block)
-
-@ExperimentalActionDSL
-@ActionDSL
-suspend inline fun ActionCoroutine.sequential(block: () -> Unit) {
-    lock()
-    block()
-    free()
-}
-
-@ExperimentalActionDSL
-class Routine(
-        val debug: Boolean,
-        val block: suspend ActionCoroutine.() -> Unit
-)
-
-@ExperimentalActionDSL
-@ActionDSL
-fun Action.runRoutine(debug: Boolean = false, block: suspend ActionCoroutine.() -> Unit) {
-    Routine(debug, block).run()
-}
-
-@ExperimentalActionDSL
-val synchronizedControl = SynchronizedControl()
