+++ src/main/kotlin/ca/warp7/frc/coroutines/ActionDSL.kt
@@ -1,0 +1,9 @@
+package ca.warp7.frc.coroutines
+
+/**
+ * [ActionDSL] marks a method to be part of the Action library
+ */
+
+@DslMarker
+@ExperimentalActionDSL
+annotation class ActionDSL
