+++ src/main/kotlin/ca/warp7/frc/coroutines/CoroutineState.kt
@@ -1,0 +1,5 @@
+package ca.warp7.frc.coroutines
+
+internal enum class CoroutineState {
+    Ready, Done
+}
