+++ src/main/kotlin/ca/warp7/frc/coroutines/CycleState.kt
@@ -1,0 +1,5 @@
+package ca.warp7.frc.coroutines
+
+internal enum class CycleState {
+    FirstCycle, Periodic, Done, Idle
+}
