+++ src/main/kotlin/ca/warp7/frc/coroutines/Routine.kt
@@ -1,0 +1,7 @@
+package ca.warp7.frc.coroutines
+
+@ExperimentalActionDSL
+class Routine(
+        val debug: Boolean,
+        val block: suspend ActionCoroutine.() -> Unit
+)
